Move `AnnotatedCompositeCodeTransformer` and `ErrorProneFork` to `refaster-support`
